task/requirements.txt
langgraph>=0.2
langchain>=0.3
langchain-anthropic>=0.3
langchain-openai>=0.3
langchain-ollama>=0.1
langchain-core>=0.3
pydantic>=2.0
python-frontmatter>=1.1
requests>=2.31
